# VMS-6927: Review fixes
<==== Mercurial data ====>
Branch: x_VMS_metadata_sdk_interface

+++ client/nx_client_desktop/src/ui/models/event_log_model.cpp
@@ -88,7 +88,7 @@ public:
 
     inline int size() const
     {
-        return m_events.size();
+        return (int) m_events.size();
     }
 
     inline vms::event::ActionData& at(int row)
